pyproject.toml
[tool.poetry]
name = "swayout"
version = "0.1.4"
description = "switch between sway output presets"
authors = ["Konstantin Koslowski <konstantin.koslowski@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.10"
i3ipc = "^2.2.1"
readchar = "^3.0.5"
pyxdg = "^0.27"
importlib-resources = "^6.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = {extras = ["toml"], version = "^6.3.2"}
pytest-html = "^3.1.1"
flake8 = "^4.0.1"
flake8-bugbear = "^22.3.23"

[tool.poetry.scripts]
swayout = "swayout.__init__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit=[".", "*/site-packages/*", "tests/*"]

[tool.coverage.report]
fail_under=75
